requirements.txt
flask==3.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.2.0
pandas
google-genai
